Add condition to error handling.
So OLS resquests do not display an error. OLS error response does not follow the pattern response.data.messages. We do nothing if the pattern is not followed. If an action needed for other services, like taxonomy, a different condition should be added. This does not avoid the CORS red warning as the error is capture after the request. If we want to avoid the call, we need to know exactly wich ontologies can be used with OLS so we will not try any one that is not supported.