Should fix the 404 problem with ols in production.
When testing locally or on wwwdev, OLS returns a status -1 for databases that are not found. In production it returns a 404 status.
I wanted the indexOf to use a service name but the error handler cannot depend on olsService as it would be a circular dependency so I am using a plain string.